# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Hello Peer是一款基于兴趣的社交应用。",
        "title": "Hello Peer API",
        "termsOfService": "API文档仅用于Hello Peer研发使用。",
        "contact": {
            "name": "Frank Sheng",
            "email": "726569998@qq.com"
        },
        "version": "0.1"
    },
    "host": "https://hellopeer.net",
    "basePath": "/api/v0.1",
    "paths": {
...
        "/user/pet/getPetDetails": {
            "get": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "description": "可获取宠物详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pet"
                ],
                "summary": "获取宠物详情",
                "operationId": "get_pet_details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/models.PetResponse"
                        }
                    }
                }
            }
        },
        "/user/pet/getPetDetailsById": {
            "get": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "description": "可根据用户id获取宠物详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pet"
                ],
                "summary": "根据用户id获取宠物详情",
                "operationId": "get_pet_details_by_id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "uid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/models.PetResponse"
                        }
                    }
                }
            }
        },
...
        "/user/pet/setPetDetails": {
            "post": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "description": "用户可设置宠物详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pet"
                ],
                "summary": "设置宠物详情",
                "operationId": "set_pet_details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "宠物详情",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PetRequest"
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.PetRequest": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string",
                    "example": "2021-10-12"
                },
                "description": {
                    "type": "string",
                    "example": "这是一个好宠物"
                },
                "pet_Name": {
                    "type": "string",
                    "example": "狗狗2022"
                },
                "sex": {
                    "type": "string",
                    "example": "MM"
                },
                "weight": {
                    "type": "number",
                    "example": 3.2
                }
            }
        },
        "models.PetResponse": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string",
                    "example": "2021-10-12"
                },
                "description": {
                    "type": "string",
                    "example": "这是一个好宠物"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "img1.jpg",
                        "img2.jpg"
                    ]
                },
                "pet_Name": {
                    "type": "string",
                    "example": "狗狗2022"
                },
                "sex": {
                    "type": "string",
                    "example": "MM"
                },
                "weight": {
                    "type": "number",
                    "example": 3.2
                }
            }
        },
...
    }
}